Organize helper functions
This change moves code into more clearly-named modules.

https://phabricator.endlessm.com/T32972
