+++ src/kolibri_daemon/application.py
@@ -16,10 +16,10 @@ from kolibri_app.config import DAEMON_PRIVATE_OBJECT_PATH
 from .dbus_helpers import DBusManagerProxy
 from .desktop_users import AccountsServiceManager
 from .desktop_users import UserInfo
+from .futures import future_chain
+from .glib_helpers import dict_to_vardict
 from .kolibri_search_handler import LocalSearchHandler
 from .kolibri_service_manager import KolibriServiceManager
-from .utils import dict_to_vardict
-from .utils import future_chain
 
 
 INACTIVITY_TIMEOUT_MS = 30 * 1000  # 30 seconds in milliseconds
